Fix idempotence of tweaks::ubuntu_service_override
To achieve that, the commit adds custom function named is_pkg_installed,
which uses dpkg-query or rpm to check the package existence. No service
override magic happens if package is already installed.

Closes-Bug: 1524327

+++ deployment/puppet/tweaks/manifests/ubuntu_service_override.pp
@@ -7,25 +7,26 @@ define tweaks::ubuntu_service_override (
     $file_name     = "create_${service_name}_override"
     $exec_name     = "remove_${service_name}_override"
 
-    file { $file_name :
-      ensure  => present,
-      path    => $override_file,
-      content => 'manual',
-      mode    => '0644',
-      owner   => 'root',
-      group   => 'root',
-    }
+    if ! is_pkg_installed($package_name) {
+      file { $file_name :
+        ensure  => present,
+        path    => $override_file,
+        content => 'manual',
+        mode    => '0644',
+        owner   => 'root',
+        group   => 'root',
+      }
 
-    exec { $exec_name :
-      path    => [ '/sbin', '/bin', '/usr/bin', '/usr/sbin' ],
-      command => "rm -f ${override_file}",
-      onlyif  => "test -f ${override_file}",
-    }
+      exec { $exec_name :
+        path    => [ '/sbin', '/bin', '/usr/bin', '/usr/sbin' ],
+        command => "rm -f ${override_file}",
+        onlyif  => "test -f ${override_file}",
+      }
 
-    File[$file_name] -> Package <| name == $package_name |> -> Exec[$exec_name]
-    File[$file_name] -> Package <| title == $package_name |> -> Exec[$exec_name]
-    File[$file_name] -> Exec[$exec_name]
-    Exec[$exec_name] -> Service <| name == $service_name |>
-    Exec[$exec_name] -> Service <| title == $service_name |>
+      File[$file_name] -> Package <| name == $package_name |> -> Exec[$exec_name]
+      File[$file_name] -> Package <| title == $package_name |> -> Exec[$exec_name]
+      Exec[$exec_name] -> Service <| name == $service_name |>
+      Exec[$exec_name] -> Service <| title == $service_name |>
+    }
   }
 }
